+++ src/main/groovy/org/codenarc/rule/unused/UnusedMethodParameterRule.groovy
@@ -43,9 +43,13 @@ class UnusedMethodParameterAstVisitor extends AbstractAstVisitor {
 
     @Override
     protected void visitConstructorOrMethod(MethodNode node, boolean isConstructor) {
-        if (!currentClassNode.isInterface() && !node.isAbstract() && !(currentClassName ==~ rule.ignoreClassRegex)
-            && !Modifier.isPrivate(node.modifiers) && AstUtil.getAnnotation(node, 'Override') == null
-            && !isMainMethod(node)) {
+        if (!currentClassNode.isInterface()
+                && !node.isAbstract()
+                && !(currentClassName ==~ rule.ignoreClassRegex)
+                && !Modifier.isPrivate(node.modifiers)
+                && AstUtil.getAnnotation(node, 'Override') == null
+                && AstUtil.getAnnotation(node, 'Pointcut') == null
+                && !isMainMethod(node)) {
             def unusedParameterNames = node.parameters*.name
             def collector = new ReferenceCollector()
             collector.visitMethod(node)